Refactored main menu and improved UI
