+++ static/src/ts/views/Profile/Replays.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import ButtonComponent from '../../components/UI/ButtonComponent';
 import UINavigator from '../../models/UINavigator';
-import GameplayOptions from '../GameplayOptions';
+import GameplayOptions from '../MainMenu';
 import Profile from './Profile';
 
 type ReplaysProps = Record<string, never>;
